+++ packages/storybook/stories/MessageThread/snippets/MessageWithAttachmentFromTeams.snippet.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 
 initializeFileTypeIcons();
 
-export const MessageWithAttachment: () => JSX.Element = () => {
+export const MessageWithAttachmentFromTeams: () => JSX.Element = () => {
   return (
     <FluentThemeProvider>
       <MessageThread
@@ -13,10 +13,10 @@ export const MessageWithAttachment: () => JSX.Element = () => {
         messages={[
           {
             messageType: 'chat',
-            senderId: '8:acs:7a7894bb-0d53-4fe8-b83e-dce84ade5e89',
-            senderDisplayName: 'Robert Tolbert',
-            messageId: `64d334be-f30c-4218-bc1c-853a28512187`,
-            content: 'Here is my router configuration file',
+            senderId: '8:orgid:f949b075-4ab2-43de-8803-e01371df20a3',
+            senderDisplayName: 'Kat Larsson',
+            messageId: `f0a6096c-da7b-42d0-9d67-05297c926bf9`,
+            content: 'Checkout these files from my OneDrive, let me know what do you think',
             createdOn: new Date('2019-04-13T00:00:00.000+08:10'),
             mine: false,
             attached: false,
@@ -24,8 +24,13 @@ export const MessageWithAttachment: () => JSX.Element = () => {
             contentType: 'text',
             attachments: [
               {
-                id: 'p121eacd612d',
-                name: 'RouterConfig.txt',
+                id: 'a021eacd612d',
+                name: 'Report.ppt',
+                url: 'https://www.onedrive.com'
+              },
+              {
+                id: 'zc1eacd612d',
+                name: 'Budget.xlsx',
                 url: 'https://www.bing.com'
               }
             ]
@@ -33,9 +38,9 @@ export const MessageWithAttachment: () => JSX.Element = () => {
           {
             messageType: 'chat',
             senderId: 'user2',
-            senderDisplayName: 'Kat Larsson',
-            messageId: `348eda2f-1582-4f29-bb5d-cda7295ca398`,
-            content: 'Thanks for sharing. This will help me troubleshoot your internet connection.',
+            senderDisplayName: 'Robert Tolbert',
+            messageId: `2a57431b-0b86-4e37-92e9-37fc03553a9d`,
+            content: 'will do, thanks!',
             createdOn: new Date('2019-04-13T00:00:00.000+08:09'),
             mine: true,
             attached: false,
